Working with python 2 and 3

+++ NFS_test/tear_down.py
@@ -3,7 +3,6 @@ Description
 """
 
 from platform import dist
-import subprocess as sp
 import logging
 import socket
 import sys
@@ -14,18 +13,6 @@ import common
 
 
 DEBUG_LOG = logging.getLogger(DEBUG_LOG)
-
-def remove_dir(directory, skip=None):
-    """ Description """
-    for entry in os.scandir(directory):
-        if entry.path == skip:
-            continue
-        elif entry.is_dir(follow_symlinks=False):
-            remove_dir(entry.path)
-        else:
-            os.remove(entry.path)
-    os.rmdir(directory)
-
 
 class Suite:
     """ Description """
@@ -41,42 +28,36 @@ class Suite:
         #
         with open(EXPORTS_PATH) as file:
             existed = file.readlines()
-        existed.remove(EXPORTS_LINE)
-        with open(EXPORTS_PATH, 'w') as file:
-            file.writelines(existed)
+        if EXPORTS_LINE in existed:
+            existed.remove(EXPORTS_LINE)
+            with open(EXPORTS_PATH, 'w') as file:
+                file.writelines(existed)
         #
-        sp.run([PACKAGE_MANAGER, 'remove', '-y'] + NFS_UTILS)
+        common.execute([PACKAGE_MANAGER, 'remove', '-y'] + NFS_UTILS)
 
     @classmethod
-    def client(cls, test_dir, log):
-        try:
-            remove_dir(test_dir, skip=ENV_DIR)
-        except OSError as error:
-            DEBUG_LOG.error(error)
-        else:
-            log.info("Removed - %s" % test_dir)
+    def client(cls, test_dir):
+        """ Description """
+        remove = common.execute(['rm', '-r', test_dir], collect=True)
         #
         for util in NFS_UTILS:
-            stop = common.execute(["service", util, "stop"], stdout=sp.PIPE, stderr=sp.PIPE)
-            common.write_to([DEBUG_LOG.debug, DEBUG_LOG.error], stop)
+            stop = common.execute(["service", util, "stop"], collect=True)
         #
         util_rem = common.execute([PACKAGE_MANAGER, 'remove', '-y'] + NFS_UTILS,
-                                  stdout=sp.PIPE, stderr=sp.PIPE)
-        common.write_to([DEBUG_LOG.debug, DEBUG_LOG.error], util_rem)
-        if not util_rem[2]: # Exit code
-            log.info("NFS removed")
+                                  collect=True)
+        return remove[2] and util_rem[2]
 
 
 class Case:
     """ Description """
 
     @classmethod
-    def client(cls, test_dir, log):
+    def client(cls, test_dir):
         """ Description """
-        unmount = common.execute(['umount', test_dir], stdout=sp.PIPE, stderr=sp.PIPE)
-        if not unmount[2]: # Exit code
-            log.info("Unmounted - %s" % test_dir)
-        common.write_to([DEBUG_LOG.debug, DEBUG_LOG.error], unmount)
+        unmount = common.execute(['umount', test_dir], collect=True)
+        if unmount[2]: # Exit code
+            DEBUG_LOG.debug("Unmounted - %s" % test_dir)
+        return unmount[2]
 
 if socket.gethostbyname(socket.gethostname()) == SERVER_ADDRESS:
     Suite.server(sys.argv.pop())
